refactor(ui): extract FieldSizeEditor into its own file
